Refactor DocumentCanvas and Header components by removing unused code and simplifying imports. Streamline the Header component to focus on sidebar functionality, enhancing overall code clarity and maintainability.
